+++ src/listen/mod.rs
@@ -1,4 +1,5 @@
 //! Listen module
 
+pub mod flux;
 pub mod rest;
 pub mod websocket;